ship: iteration 4 — preserve project routes from Chats, settle empty chats (codex P2s)
Project-tab clicks from the Chats machine tab now navigate to the
current binding's stored route (extracted read/write helpers into
app/projectRouteStorage.ts, shared by ProjectTabHost and TopBar) so
returning from Chats no longer stamps /work over a project's remembered
/prs//files position. A selected chat whose history has settled empty
(e.g. failed first send) renders a stable empty state instead of an
indefinite spinner.
