+++ apps/desktop/src/renderer/components/app/projectRouteStorage.ts
@@ -1,0 +1,26 @@
+/// Per-surface route memory: each open project tab remembers its last route
+/// under `ade:project-route:<bindingKey>` so tab switches replay where the
+/// user left off. Shared by ProjectTabHost (read/write on tab swaps) and
+/// TopBar (read when routing back to a project from the Chats machine tab).
+const PROJECT_ROUTE_STORAGE_PREFIX = "ade:project-route:";
+
+function projectRouteStorageKey(projectRoot: string): string {
+  return `${PROJECT_ROUTE_STORAGE_PREFIX}${projectRoot}`;
+}
+
+export function readStoredProjectRoute(projectRoot: string): string | null {
+  try {
+    const value = window.localStorage.getItem(projectRouteStorageKey(projectRoot));
+    return value?.startsWith("/") ? value : null;
+  } catch {
+    return null;
+  }
+}
+
+export function writeStoredProjectRoute(projectRoot: string, route: string): void {
+  try {
+    window.localStorage.setItem(projectRouteStorageKey(projectRoot), route);
+  } catch {
+    // localStorage can be unavailable in private/test environments.
+  }
+}
